Refactor a little RestJson1 and add simple RFC doc
