+++ codegen-server/src/main/kotlin/software/amazon/smithy/rust/codegen/server/smithy/generators/ServiceGenerator.kt
@@ -5,25 +5,24 @@
 
 package software.amazon.smithy.rust.codegen.server.smithy.generators
 
-import java.util.logging.Logger
 import software.amazon.smithy.model.knowledge.TopDownIndex
-import software.amazon.smithy.rust.codegen.smithy.generators.HttpProtocolGenerator
-import software.amazon.smithy.rust.codegen.smithy.generators.ProtocolSupport
-import software.amazon.smithy.rust.codegen.smithy.generators.ProtocolConfig
 import software.amazon.smithy.rust.codegen.rustlang.RustModule
 import software.amazon.smithy.rust.codegen.smithy.RustCrate
 import software.amazon.smithy.rust.codegen.smithy.customize.RustCodegenDecorator
+import software.amazon.smithy.rust.codegen.smithy.generators.HttpProtocolGenerator
+import software.amazon.smithy.rust.codegen.smithy.generators.ProtocolConfig
+import software.amazon.smithy.rust.codegen.smithy.generators.ProtocolSupport
 import software.amazon.smithy.rust.codegen.smithy.generators.config.ServiceConfigGenerator
 import software.amazon.smithy.rust.codegen.smithy.generators.error.CombinedErrorGenerator
 import software.amazon.smithy.rust.codegen.smithy.generators.error.TopLevelErrorGenerator
 import software.amazon.smithy.rust.codegen.util.inputShape
 
 class ServiceGenerator(
-    private val rustCrate: RustCrate,
-    private val protocolGenerator: HttpProtocolGenerator,
-    private val protocolSupport: ProtocolSupport,
-    private val config: ProtocolConfig,
-    private val decorator: RustCodegenDecorator,
+        private val rustCrate: RustCrate,
+        private val protocolGenerator: HttpProtocolGenerator,
+        private val protocolSupport: ProtocolSupport,
+        private val config: ProtocolConfig,
+        private val decorator: RustCodegenDecorator,
 ) {
     private val index = TopDownIndex.of(config.model)
 
@@ -33,15 +32,16 @@ class ServiceGenerator(
             rustCrate.useShapeWriter(operation) { operationWriter ->
                 rustCrate.useShapeWriter(operation.inputShape(config.model)) { inputWriter ->
                     protocolGenerator.renderOperation(
-                        operationWriter,
-                        inputWriter,
-                        operation,
-                        decorator.operationCustomizations(config, operation, listOf())
+                            operationWriter,
+                            inputWriter,
+                            operation,
+                            decorator.operationCustomizations(config, operation, listOf())
                     )
                 }
             }
             rustCrate.withModule(RustModule.Error) { writer ->
-                CombinedErrorGenerator(config.model, config.symbolProvider, operation).render(writer)
+                CombinedErrorGenerator(config.model, config.symbolProvider, operation)
+                        .render(writer)
             }
         }
 
@@ -49,14 +49,12 @@ class ServiceGenerator(
 
         rustCrate.withModule(RustModule.Config) { writer ->
             ServiceConfigGenerator.withBaseBehavior(
-                config,
-                extraCustomizations = decorator.configCustomizations(config, listOf())
-            ).render(writer)
+                            config,
+                            extraCustomizations = decorator.configCustomizations(config, listOf())
+                    )
+                    .render(writer)
         }
 
-        rustCrate.lib {
-            it.write("pub use config::Config;")
-        }
+        rustCrate.lib { it.write("pub use config::Config;") }
     }
 }
-
