+++ codegen-test/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
 data class CodegenTest(val service: String, val module: String, val extraConfig: String? = null)
 
 val CodegenTests = listOf(
+    CodegenTest("com.amazonaws.simple#SimpleService", "simple"),
     CodegenTest("com.amazonaws.dynamodb#DynamoDB_20120810", "dynamo"),
     CodegenTest("com.amazonaws.ebs#Ebs", "ebs"),
     CodegenTest("aws.protocoltests.json10#JsonRpc10", "json_rpc10"),
