+++ design/src/SUMMARY.md
@@ -16,3 +16,4 @@
   - [RFC-0001: Sharing configuration between multiple clients](./rfcs/rfc0001_shared_config.md)
   - [RFC-0002: Supporting multiple HTTP versions for SDKs that use Event Stream](./rfcs/rfc0002_http_versions.md)
   - [RFC-0003: API for Pre-signed URLs](./rfcs/rfc0003_presigning_api.md)
+  - [RFC-0004: Smithy Rust service framework](./rfcs/rfc0004_service_generation.md)
